fix(shortcuts): address review findings across shortcuts feature
- Extract shared formatKeyPart/isMac into src/lib/shortcut-format.ts,
  removing duplication from KeyRecorder and ShortcutRow
- Fix useEffect leak in KeyRecorder by adding stop() cleanup on unmount
- Fix conflict resolution logic to unbind default-key conflicts via
  empty string sentinel instead of no-op delete
- Add try/catch error handling to all updateKeyboardShortcuts mutations
- Throw descriptive error instead of silent return when settings missing
- Replace HashMap<String, serde_json::Value> with typed ShortcutKey enum
- Use @/ path aliases consistently for component imports
- Fix bg-background -> bg-card on KeyRecorder container
- Fix planning docs: correct shortcut counts, remove absolute paths,
  sync ROADMAP checklist with execution status
