+++ .planning/phases/27-settings-page/27-01-PLAN.md
@@ -23,7 +23,7 @@ requirements:
 must_haves:
   truths:
     - 'Shortcuts category appears in Settings sidebar between Appearance and Sync'
-    - 'All ~10 shortcut definitions are active (not commented out)'
+    - 'All 9 shortcut definitions are active (not commented out)'
     - 'keyboard_shortcuts field exists in both Rust and TypeScript Settings types with serde(default)'
     - 'Shortcuts are displayed grouped by scope with section headers'
   artifacts:
@@ -67,12 +67,12 @@ must_haves:
 Add the keyboard_shortcuts field to backend/frontend Settings types, activate all commented-out shortcut definitions, register the Shortcuts settings category, and create the read-only shortcut list UI grouped by scope.
 
 Purpose: Establish the data model and display layer so Plan 02 can add interactivity (key recording, conflict detection, persistence).
-Output: Shortcuts section visible in Settings with all ~10 shortcuts displayed by scope, keyboard_shortcuts field available for persistence.
+Output: Shortcuts section visible in Settings with all 9 shortcuts displayed by scope, keyboard_shortcuts field available for persistence.
 </objective>
 
 <execution_context>
-@/home/wuy6/.claude/get-shit-done/workflows/execute-plan.md
-@/home/wuy6/.claude/get-shit-done/templates/summary.md
+@.claude/get-shit-done/workflows/execute-plan.md
+@.claude/get-shit-done/templates/summary.md
 </execution_context>
 
 <context>
@@ -210,16 +210,16 @@ const { setting, loading, updateSetting } = useSetting()
      - search.focus (cmd+/, scope: global)
      - modal.close (esc, scope: modal)
    - Use `mod` prefix instead of `cmd` for cross-platform compatibility (mod = cmd on Mac, ctrl on Windows/Linux). Check existing active shortcuts for the convention -- `nav.settings` uses `mod+comma`, so follow that.
-   - Keep the clipboard.copy shortcut commented if it was commented (the research mentions ~8 commented shortcuts, activate all that exist)
+   - Note: clipboard.copy exists in the ShortcutAction type but has no definition entry -- do not create one (it is not a commented-out shortcut, it simply has no definition)
      </action>
      <verify>
-     <automated>cd /home/wuy6/myprojects/UniClipboard/src-tauri && cargo test --lib -p uc-core settings::model::tests -- --include-ignored 2>&1 | tail -20</automated>
+     <automated>cd src-tauri && cargo test --lib -p uc-core settings::model::tests -- --include-ignored 2>&1 | tail -20</automated>
      </verify>
      <done>
    - Rust Settings struct has keyboard_shortcuts: HashMap with serde(default)
    - TypeScript Settings interface has keyboard_shortcuts optional field
    - SettingContext exposes updateKeyboardShortcuts helper
-   - All ~8 commented shortcut definitions are active in SHORTCUT_DEFINITIONS array
+   - All 7 previously-commented shortcut definitions are active in SHORTCUT_DEFINITIONS array (9 total)
    - Existing cargo tests pass including new backward-compat test
      </done>
      </task>
@@ -317,11 +317,11 @@ const { setting, loading, updateSetting } = useSetting()
 
      </action>
      <verify>
-       <automated>cd /home/wuy6/myprojects/UniClipboard && npx tsc --noEmit 2>&1 | tail -20</automated>
+       <automated>npx tsc --noEmit 2>&1 | tail -20</automated>
      </verify>
      <done>
        - Shortcuts category appears in Settings sidebar between Appearance and Sync with Command icon
-       - ShortcutsSection renders all ~10 shortcuts grouped by scope
+       - ShortcutsSection renders all 9 shortcuts grouped by scope
        - Each row shows description, key badge with platform-appropriate symbols, edit button
        - Reset button only visible on modified shortcuts
        - Reset All button at bottom, disabled when no overrides
@@ -341,7 +341,7 @@ const { setting, loading, updateSetting } = useSetting()
 <success_criteria>
 
 - Shortcuts settings category visible and functional in Settings sidebar
-- All ~10 shortcut definitions active (not commented out)
+- All 9 shortcut definitions active (not commented out)
 - keyboard_shortcuts field available in both Rust and TypeScript Settings types
 - Shortcuts displayed with proper grouping, key badges, and i18n
 - No compilation errors in Rust or TypeScript
