+++ .planning/phases/27-settings-page/27-02-PLAN.md
@@ -64,8 +64,8 @@ Output: Fully interactive Shortcuts settings section with immediate effect on ac
 </objective>
 
 <execution_context>
-@/home/wuy6/.claude/get-shit-done/workflows/execute-plan.md
-@/home/wuy6/.claude/get-shit-done/templates/summary.md
+@.claude/get-shit-done/workflows/execute-plan.md
+@.claude/get-shit-done/templates/summary.md
 </execution_context>
 
 <context>
@@ -179,7 +179,7 @@ const [keys, { start, stop, isRecording }] = useRecordHotkeys()
    - "Reset All Shortcuts" button enabled when `Object.keys(setting?.keyboard_shortcuts ?? {}).length > 0`
      </action>
      <verify>
-     <automated>cd /home/wuy6/myprojects/UniClipboard && npx tsc --noEmit 2>&1 | tail -20</automated>
+     <automated>npx tsc --noEmit 2>&1 | tail -20</automated>
      </verify>
      <done>
    - KeyRecorder captures key combos using useRecordHotkeys
@@ -221,7 +221,7 @@ const [keys, { start, stop, isRecording }] = useRecordHotkeys()
    - Do NOT touch call sites for the newly-uncommented shortcuts until those shortcuts are actually wired to handlers (which is out of scope for this plan -- the definitions are active but handlers are wired in future work)
      </action>
      <verify>
-     <automated>cd /home/wuy6/myprojects/UniClipboard && npx tsc --noEmit 2>&1 | tail -20</automated>
+     <automated>npx tsc --noEmit 2>&1 | tail -20</automated>
      </verify>
      <done>
    - useShortcut hook accepts optional `id` parameter
@@ -240,7 +240,7 @@ const [keys, { start, stop, isRecording }] = useRecordHotkeys()
     1. Run `bun tauri dev`
     2. Open Settings (Cmd+Comma or click gear icon)
     3. Verify "Shortcuts" category appears in sidebar between "Appearance" and "Sync" with Command icon
-    4. Click "Shortcuts" -- verify all ~10 shortcuts displayed grouped by scope (Global, Clipboard, Settings, Devices, Modal)
+    4. Click "Shortcuts" -- verify all 9 shortcuts displayed grouped by scope (Global, Clipboard, Settings, Devices, Modal)
     5. Click edit (pencil) on any shortcut -- verify recording mode activates (visual indicator)
     6. Press a new key combo -- verify it displays as key badges
     7. Press Escape during recording -- verify it cancels (no change)
